+++ functions/src/shared.ts
@@ -68,21 +68,6 @@ export const getCurriculumMetadata = (sectionPath?: string): ICurriculumMetadata
           : undefined;
 }
 
-export function buildFirebaseImageUrl(classHash: string, imageKey: string) {
-  return `ccimg://fbrtdb.concord.org/${classHash}/${imageKey}`;
-}
-
-/*
- * Parses a firebase image url into its constituent parts.
- */
-export function parseFirebaseImageUrl(url: string) {
-  const match = /ccimg:\/\/fbrtdb\.concord\.org\/([^/]+)(\/([^/]+))?/.exec(url);
-  const imageKey = match?.[3] || match?.[1];
-  const imageClassHash = match?.[3] ? match?.[1] : undefined;
-  const legacyUrl = imageClassHash ? url.replace(`/${imageClassHash}`, ""): url;
-  return { imageClassHash, imageKey, legacyUrl };
-}
-
 /*
  * Types that are shared between cloud functions and client code.
  */
